Sharing content on website chapter 5
Create bookmarking website and make functionality to post content
from other websites, create a detail view for the images and use
thumbnail.

And implement Ajax actions with jQuery, create custom decorator to image_like view
to check it is a ajax request, then add ajax pagination to the list view.

-Note FB and Twitter social authentication is difficult to make.